+++ CHANGES.md
@@ -1,7 +1,12 @@
 ### 0.9.8
 
+* Constraint the string tags to contain only alpha-numeric characters
+  and few mores (`-`, `_` and `/`) (#186)
+* Fix a race condition in `Irmin.clone`. (#221)
 * Escpate double quotes in the output of commit messages to workaround
-  HTML display issues (#222)
+  HTML display issues. (#222)
+* Update to the latest version of dagre-d3 to create nicely animated
+  graphs of the store.
 
 ### 0.9.7 (2015-07-06)